requirements.txt
transformers
sentence-transformers
faiss-cpu
PyPDF2
PyYAML
numpy